% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_poly.R
\name{eval_poly}
\alias{eval_poly}
\title{Evaluate a \code{poly} onbject}
\usage{
eval_poly(poly, x)
}
\arguments{
\item{poly}{\code{poly} object}

\item{x}{A vector of values}
}
\value{
A vector of same length of \code{x}, of \code{poly} evaluated at
  values of \code{x}.
}
\description{
Evaluate a \code{poly} onbject
}
