% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\name{make_poly}
\alias{make_poly}
\title{Create a \code{poly} object}
\usage{
make_poly(...)
}
\arguments{
\item{...}{Named arguments where the name is the power, and the value is its
corresponding coefficient. E.g. \code{"2" = 5} implies $5x^2$. If multiple
arguments have the same power (e.g. \code{make_poly("2" = 1, "2" = 5)}),
the coefficients will be summed (continuning example, $x^2 + 5x^2 =
6x^2$).}
}
\value{
A new \code{poly} object
}
\description{
Create a \code{poly} object
}
