% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\docType{class}
\name{poly-class}
\alias{poly-class}
\title{A class represnting a polynomial equation}
\description{
A class represnting a polynomial equation
}
\section{Slots}{

\describe{
\item{\code{The}}{powers crresoponding to each coefficient}
}}

