% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.poly.R
\name{plot.poly}
\alias{plot.poly}
\alias{ggplot_poly}
\title{Plot a \code{poly} object}
\usage{
\method{plot}{poly}(x, y = 0, to = 1, from = y, ...)

ggplot_poly(poly, from = 0, to = 1, ...)
}
\arguments{
\item{x, poly}{a \code{poly} object}

\item{y}{alias for \code{from} for compatibility with \code{plot}}

\item{to}{right side of range to plot}

\item{from}{left side of range to plot}
}
\value{
The plot
}
\description{
Plot a \code{poly} object
}
