% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_numeric_math.R
\name{*,poly,numeric-method}
\alias{*,poly,numeric-method}
\alias{*,numeric,poly-method}
\alias{/,poly,numeric-method}
\alias{/,numeric,poly-method}
\alias{+,poly,numeric-method}
\alias{+,numeric,poly-method}
\alias{-,poly,numeric-method}
\alias{-,numeric,poly-method}
\title{\code{poly} and numeric arithmetic

poly/numeric is not supported. Other operations are supported. Exactly one
of \code{e1} or \code{e2} should be \code{poly}.}
\usage{
\S4method{*}{poly,numeric}(e1, e2)

\S4method{*}{numeric,poly}(e1, e2)

\S4method{/}{poly,numeric}(e1, e2)

\S4method{/}{numeric,poly}(e1, e2)

\S4method{+}{poly,numeric}(e1, e2)

\S4method{+}{numeric,poly}(e1, e2)

\S4method{-}{poly,numeric}(e1, e2)

\S4method{-}{numeric,poly}(e1, e2)
}
\arguments{
\item{e1}{\code{poly} or \code{numeric}.}

\item{e2}{\code{poly} or \code{numeric}.}
}
\value{
\code{poly}
}
\description{
\code{poly} and numeric arithmetic

poly/numeric is not supported. Other operations are supported. Exactly one
of \code{e1} or \code{e2} should be \code{poly}.
}
