% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_poly_math.R
\name{+,poly,poly-method}
\alias{+,poly,poly-method}
\alias{-,poly,poly-method}
\alias{*,poly,poly-method}
\alias{/,poly,poly-method}
\title{\code{poly}-\code{poly} math}
\usage{
\S4method{+}{poly,poly}(e1, e2)

\S4method{-}{poly,poly}(e1, e2)

\S4method{*}{poly,poly}(e1, e2)

\S4method{/}{poly,poly}(e1, e2)
}
\arguments{
\item{e1}{\code{poly} object}

\item{e2}{\code{poly} object}
}
\value{
\code{poly} object
}
\description{
\code{poly}-\code{poly} math
}
\details{
Supports `+`, `-`, and `*`. Division not supported.
}
