% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polydat}
\alias{polydat}
\title{Polynomial data}
\format{
A data frame for using with polynomial
\describe{
  \item{a}{Random normal variable}
  \item{b}{Draw from 1-100}
}
}
\source{
Made up
}
\usage{
polydat
}
\description{
Polynomial data
}
\keyword{datasets}
