% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\name{show,poly-method}
\alias{show,poly-method}
\title{Show a \code{poly} object}
\usage{
\S4method{show}{poly}(object)
}
\arguments{
\item{object}{A \code{poly} object}
}
\value{
\code{object}, invisibly.
}
\description{
Show a \code{poly} object
}
