% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_frequencia.R
\name{tabela_frequencia}
\alias{tabela_frequencia}
\title{Tabela de Distribuição de Frequência}
\usage{
tabela_frequencia(variavel_disc_cont)
}
\arguments{
\item{variavel_disc_cont}{é uma coluna de um dataframe}
}
\value{
a tabela de distribuição de frequência da \code{variavel_disc_cont}
}
\description{
\code{tabela_frequencia()} retorna a tabela de distribuição de frequência
da \code{variavel_disc_cont}.
}
\examples{
# Tabela de frequência 4/2 = 2
tabela_frequencia(mb$Estatist.)

}
